requirements.txt
# Core ML (pinned for compatibility)
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
diffusers==0.30.3
transformers==4.41.2
accelerate==0.34.2
peft==0.12.0
safetensors>=0.4.3
huggingface_hub>=0.23.0

# CLIP & Vision
open_clip_torch>=2.23.0
Pillow>=10.0.0

# Image Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scikit-image>=0.22.0
scipy>=1.11.0

# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
aiofiles>=22.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utils
numpy>=1.24.0
tqdm>=4.66.0
